feat(api): clean api interfaces
