+++ src/telegram/takeover/alert.rs
@@ -1,5 +1,5 @@
 use crate::{
-    api::{HttpClient, SolRpcApi},
+    api::{DexScreenerApi, PumpApi, SolRpcApi},
     config,
     context::{Context, TaskCache},
     model::{Alert, AlertTitle},
